requirements.txt
firebase-admin
google-cloud-storage
python-dotenv
PyQt5==5.15.9
qtawesome
requests
pyaudio
aiortc
av
aiohttp
qasync